+++ RELEASE_NOTES.md
@@ -1,4 +1,8 @@
 ## [4.0 Snapshot 3] - UNRELEASED
+### New Features
+#### Advancements
+- Added new "Traveler" adventure challenge for sleeping in a tent at least 1 km away from spawn... or in a different dimension
+
 ### Changes
 #### Blast Fungus
 - Added new particle effect
@@ -11,6 +15,9 @@
 ### Technical Changes
 #### Particles
 - Added `spore_cloud` particle
+
+#### Stats
+- Added `Times Slept in a Tent` statistic
 
 #### Bug Fixes
 - Fixed bug that prevented enchanted books from using their vanilla enchantment costs to calculate Magic Capacity